requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
django-model-utils==4.3.1
python-json-logger==2.0.4
django-extensions==3.2.1
psycopg2-binary==2.9.5
gunicorn==20.1.0
